requirements.txt
# Core dependencies
requests>=2.28.2
beautifulsoup4>=4.11.2
loguru>=0.6.0
python-dotenv>=1.0.0
selenium>=4.8.0
webdriver-manager>=3.8.5

# Data processing
pandas>=1.5.3
numpy>=1.24.2

# Database
pymongo>=4.3.3
neo4j>=5.6.0

# Graph analysis
networkx>=3.0

# NLP and ML
scikit-learn>=1.2.1
nltk>=3.8.1

# HTML parsing
lxml>=4.9.2
html5lib>=1.1

# Image processing
Pillow>=9.4.0

# Testing
pytest>=7.2.2

# Utils
tqdm>=4.65.0
colorama>=0.4.6
